Improve updates section scraper and update DOCS.md
The updates section now contains not only <p> tags, but also <ul> tags. This part of the scraper now iterates over the children tags of the update section <div> in order to write all update text to the README. As such, a more constricting set of CSS classes is now used to avoid searching children that are just other containing <div>s, and documentation was modified to reflect the change.